facedetection/requirement.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asgiref==3.9.1
astunparse==1.6.3
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.3
Django==5.2.5
djangorestframework==3.16.1
easydict==1.13
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface-rk==0.7.3
joblib==1.5.2
keras==3.11.3
keras-facenet==0.3.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.5
numpy==2.1.3
onnx==1.19.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
retina-face==0.0.17
rich==14.1.0
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
simsimd==6.5.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
stringzilla==3.12.6
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2025.8.28
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3